Add files via upload
This project uses two files that should be included here, but in the case of the memory, one of the files is not uploaded for the project. This code creates a new database and loading data to it. Then it uses the relations between tables and takes the output of the top5 played tracks and top played artist.